+++ packages/plugins/type-designer/src/headless/constants/type-elements.ts
@@ -7,7 +7,6 @@ export const TYPE_FAMILY_MAP = {
 } as const
 
 export const REF_FAMILY_MAP = {
-	bay: 'bay',
 	generalEquipment: 'generalEquipment',
 	conductingEquipment: 'conductingEquipment',
 	function: 'function',
@@ -22,22 +21,15 @@ export const COLUMNS = {
 	lNodeType: 'lNodeType'
 } as const
 
-export const REF_FAMILY_TO_TYPE_FAMILY_MAP = {
-	[REF_FAMILY_MAP.bay]: REF_FAMILY_MAP.bay,
-	[REF_FAMILY_MAP.generalEquipment]: TYPE_FAMILY_MAP.generalEquipmentType,
-	[REF_FAMILY_MAP.conductingEquipment]:
-		TYPE_FAMILY_MAP.conductingEquipmentType,
-	[REF_FAMILY_MAP.function]: TYPE_FAMILY_MAP.functionTemplate,
-	[REF_FAMILY_MAP.eqFunction]: TYPE_FAMILY_MAP.functionTemplate,
-	[REF_FAMILY_MAP.lNode]: TYPE_FAMILY_MAP.lNodeType
-}
-
-export const TYPE_FAMILY_EQUIVALENT_FOR_ATTRIBUTES = {
-	...TYPE_FAMILY_MAP,
-	// standard equivalent for custom elements
-	generalEquipmentType: 'generalEquipment',
-	conductingEquipmentType: 'conductingEquipment'
+export const KIND = {
+	custom: 'custom',
+	standard: 'standard',
+	unstable: 'unstable'
 } as const
+
+export const READONLY_ATTRIBUTES = ['id', 'uuid', 'originUuid'] as const
+
+//====== MAPPERS ======//
 
 export const TYPE_FAMILY_TO_COLUMN_KEY = {
 	[TYPE_FAMILY_MAP.bay]: COLUMNS.bay,
@@ -47,22 +39,53 @@ export const TYPE_FAMILY_TO_COLUMN_KEY = {
 	[TYPE_FAMILY_MAP.lNodeType]: COLUMNS.lNodeType
 } as const
 
+export const TYPE_FAMILY_TO_REF_FAMILY_MAP = {
+	[TYPE_FAMILY_MAP.generalEquipmentType]: REF_FAMILY_MAP.generalEquipment,
+	[TYPE_FAMILY_MAP.conductingEquipmentType]:
+		REF_FAMILY_MAP.conductingEquipment,
+	[TYPE_FAMILY_MAP.functionTemplate]: [
+		REF_FAMILY_MAP.function,
+		REF_FAMILY_MAP.eqFunction
+	],
+	[TYPE_FAMILY_MAP.lNodeType]: REF_FAMILY_MAP.lNode
+} as const
+
+export const REF_FAMILY_TO_TYPE_FAMILY_MAP = {
+	[REF_FAMILY_MAP.generalEquipment]: TYPE_FAMILY_MAP.generalEquipmentType,
+	[REF_FAMILY_MAP.conductingEquipment]:
+		TYPE_FAMILY_MAP.conductingEquipmentType,
+	[REF_FAMILY_MAP.function]: TYPE_FAMILY_MAP.functionTemplate,
+	[REF_FAMILY_MAP.eqFunction]: TYPE_FAMILY_MAP.functionTemplate,
+	[REF_FAMILY_MAP.lNode]: TYPE_FAMILY_MAP.lNodeType
+} as const
+
+export const TYPE_FAMILY_EQUIVALENT_FOR_ATTRIBUTES = {
+	...TYPE_FAMILY_MAP,
+	// standard equivalent for custom elements
+	[TYPE_FAMILY_MAP.generalEquipmentType]: 'generalEquipment',
+	[TYPE_FAMILY_MAP.conductingEquipmentType]: 'conductingEquipment'
+} as const
+
+export const REF_ATTRIBUTES_KIND_BY_REF_FAMILY = {
+	[REF_FAMILY_MAP.generalEquipment]: KIND.custom,
+	[REF_FAMILY_MAP.conductingEquipment]: KIND.custom,
+	[REF_FAMILY_MAP.function]: KIND.standard,
+	[REF_FAMILY_MAP.eqFunction]: KIND.standard,
+	[REF_FAMILY_MAP.lNode]: KIND.standard
+} as const
+
 export const CUSTOM_TAG_NAME_MAP = {
-	generalEquipmentType: 'GeneralEquipmentType',
-	conductingEquipmentType: 'ConductingEquipmentType'
+	[TYPE_FAMILY_MAP.generalEquipmentType]: 'GeneralEquipmentType',
+	[TYPE_FAMILY_MAP.conductingEquipmentType]: 'ConductingEquipmentType'
 } as const
 
 export const ALLOWED_TARGETS_BY_REF_FAMILY = {
-	bay: [],
-	generalEquipment: [TYPE_FAMILY_MAP.bay],
-	conductingEquipment: [TYPE_FAMILY_MAP.bay],
-	function: [
-		TYPE_FAMILY_MAP.bay,
+	[REF_FAMILY_MAP.generalEquipment]: [TYPE_FAMILY_MAP.bay],
+	[REF_FAMILY_MAP.conductingEquipment]: [TYPE_FAMILY_MAP.bay],
+	[REF_FAMILY_MAP.function]: [TYPE_FAMILY_MAP.bay],
+	[REF_FAMILY_MAP.eqFunction]: [
 		TYPE_FAMILY_MAP.conductingEquipmentType,
 		TYPE_FAMILY_MAP.generalEquipmentType
 	],
-	eqFunction: [TYPE_FAMILY_MAP.generalEquipmentType],
-	lNode: [TYPE_FAMILY_MAP.functionTemplate]
+	[REF_FAMILY_MAP.lNode]: [TYPE_FAMILY_MAP.functionTemplate]
 } as const
-
-export const READONLY_ATTRIBUTES = ['id', 'uuid', 'originUuid'] as const