+++ packages/plugins/type-designer/vite.config.ts
@@ -1,7 +1,6 @@
 import { defineConfig } from 'vite'
 import { svelte } from '@sveltejs/vite-plugin-svelte'
 import { resolve } from 'node:path'
-import jsonPackage from './package.json'
 import dts from 'vite-plugin-dts'
 
 const isDevelopment = process.env.NODE_ENV === 'development'
@@ -18,7 +17,7 @@ export default defineConfig({
 		lib: {
 			entry: resolve(__dirname, 'src/plugin.ts'),
 			formats: ['es'],
-			fileName: 'plugin'
+			fileName: 'index'
 		},
 		sourcemap: isDevelopment ? 'inline' : false
 	},